change redirect, add btn shareLink

+++ app/Actions/Jetstream/DeleteUser.php
@@ -13,7 +13,9 @@ class DeleteUser implements DeletesUsers
     public function delete(User $user): void
     {
         $user->deleteProfilePhoto();
-        $user->tokens->each->delete();
+        if (!$user->tokens == null) {
+            $user->tokens->each->delete();
+        }
         $user->delete();
     }
 }
